requirements.txt
torch==1.9.1+cu111
numpy
modAL
sklearn
skorch
blitz-bayesian-pytorch